+++ cogs/fun.py
@@ -1,7 +1,9 @@
 from discord.ext import commands
 import discord
 import random
+import json
 lolpwd = 'samples/'
+LASTWRDFILE = "data/lastword.json"
 
 
 class Fun(commands.Cog):
@@ -25,7 +27,11 @@ class Fun(commands.Cog):
 
     @commands.command(name='cough', help="Simulate cough. :)")
     async def cough(self, ctx):
-        lolcough = ["What? You being infected coronavirus?", str(self.bot.get_emoji(684291327818596362)), "Please don't:\nSneeze on me;\nCough on me;\nTalk to me,\nNo oh oh!", "🤢", "Run, run, until it's done, done, until the sun comes up in the morn'."]
+        lolcough = [
+            "What? You being infected coronavirus?", str(self.bot.get_emoji(684291327818596362)),
+            "Please don't:\nSneeze on me;\nCough on me;\nTalk to me,\nNo oh oh!",
+            "🤢", "Run, run, until it's done, done, until the sun comes up in the morn'."
+        ]
         response = random.choice(lolcough)
         msg = await ctx.send(response)
         await msg.add_reaction('👀')
@@ -33,9 +39,11 @@ class Fun(commands.Cog):
 
     @commands.command(name='test', help="Respond with test messages!")
     async def test(self, ctx):
-        loltest = ["!urban MEE6", "Am I a joke to you?", "!8ball Siriu-smart?", f"What? Are you a developer?{ctx.message.author.mention}",
-                   "Didn't expect anyone would use this command, but there it is!", "No test.", "Ping Pong!", "No.", "?????", "Siriusly, What did you expect?",
-                   "Stop.", "!8ball are you stupid?", "Vincidiot"]
+        loltest = [
+            "!urban MEE6", "Am I a joke to you?", "!8ball Siriu-smart?", f"What? Are you a developer?{ctx.message.author.mention}",
+            "Didn't expect anyone would use this command, but there it is!", "No test.", "Ping Pong!", "No.", "?????",
+            "Siriusly, What did you expect?", "Stop.", "!8ball are you stupid?", "Vincidiot"
+        ]
         response = random.choice(loltest)
         msg = await ctx.send(response)
         await msg.add_reaction('👍')
@@ -43,16 +51,25 @@ class Fun(commands.Cog):
 
     @commands.command(name='stupid', help='Shout at stupid things')
     async def stupid(self, ctx, *, args='that'):
-        await ctx.send(random.choice([f'{args} is so so stupid!', f"I can't believe how stupid {args} is!", f"Seriously, {args}'s the stupidest thing I've ever heard!", f"STUPID STUPID STUPID STUPID STUPID STUPID STUPID STUPID {args}!", f"I can't believe {args}'s even a thing."]))
+        await ctx.send(random.choice([
+            f'{args} is so so stupid!', f"I can't believe how stupid {args} is!",
+            f"Seriously, {args}'s the stupidest thing I've ever heard!",
+            f"STUPID STUPID STUPID STUPID STUPID STUPID STUPID STUPID {args}!",
+            f"I can't believe {args}'s even a thing."
+        ]))
 
     @commands.command(name='whatis', help='Tells you what the input is. /whatis minecraft')
     async def whatis(self, ctx, *, args=""):
         if args == "":
             await ctx.send("Bruh, where's the argument???")
             return
-        await ctx.send(random.choice([f"{args} is generally {args}.", f"Technically, {args} is {args}!", f"To know what {args} is, please run `!urban {args}`",
-                                      f"Well, not in a nutshell, {args} as {args} is {args} in {args} on {args} at {args} from {args} to {args}...It's just...{args}!!!!!",
-                                      f"You are so dumb that you even don't know what {args} is!"]))
+        await ctx.send(
+            random.choice([
+                f"{args} is generally {args}.", f"Technically, {args} is {args}!", f"To know what {args} is, please run `!urban {args}`",
+                f"Well, not in a nutshell, {args} as {args} is {args} in {args} on {args} at {args} from {args} to {args}...It's just...{args}!!!!!",
+                f"You are so dumb that you even don't know what {args} is!"
+            ])
+        )
 
     @commands.command(name='boomer', help="/boomer [person]", aliases=['boom', 'okboomer'])
     async def boomer(self, ctx, *, person=""):
@@ -65,7 +82,7 @@ class Fun(commands.Cog):
     @commands.group(name='media', help='/media [sub-commands]', aliases=['sent'])
     async def media(self, ctx):
         if ctx.invoked_subcommand is None:
-            await ctx.send(f"2 bed idk wat u r toking 'bout, but wut?")
+            await ctx.send("2 bed idk wat u r toking 'bout, but wut?")
             return
 
     @media.command(name='no', help='/media no, will give you file related to "no"', aliases=['nope', 'nah', 'np'])
@@ -110,6 +127,15 @@ class Fun(commands.Cog):
         await ctx.send(args)
         await ctx.message.delete()
         return
+    
+    @commands.command(name='quote', help='last message of that guy')
+    async def quote(self, ctx, members: commands.Greedy[discord.Member]):
+        lastword = json.load(open(LASTWRDFILE, 'r'))
+        result = ''
+        for member in members:
+            lastmsg = await discord.TextChannel.fetch_message(ctx.message.channel, lastword[f'g{ctx.guild.id}'][member.id])
+            result += f'[Last message]({lastmsg}) by {member.mention}:\n>>> {lastmsg.content}\n\n'
+        return await ctx.send(result)
 
 
 def setup(bot):